Refactor config loading (#7622)
